whales_be_service/pyproject.toml
[tool.poetry]
name = "whales-identification"
version = "0.1.0"
description = "Библиотека для идентификации морских млекопитающих"
authors = [
    "Baltsat Konstantin",
    "Tarasov Artem",
    "Vandanov Sergey",
    "Serov Alexandr"
]
license = "MIT"
readme = "README.md"
packages = [{include = "whales_be_service", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
absl-py = "2.1.0"
albucore = "0.0.17"
albumentations = "1.4.18"
altair = "5.4.1"
annotated-types = "0.7.0"
appnope = "0.1.4"
asttokens = "2.4.1"
astunparse = "1.6.3"
attrs = "24.2.0"
backcall = "0.2.0"
blinker = "1.8.2"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.2.2"
debugpy = "1.8.7"
decorator = "5.1.1"
docker-pycreds = "0.4.0"
eval-type-backport = "0.2.0"
exceptiongroup = "1.2.2"
executing = "2.1.0"
filelock = "3.16.1"
flatbuffers = "24.3.25"
fsspec = "2024.9.0"
gast = "0.6.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
google-auth = "2.35.0"
google-auth-oauthlib = "1.2.1"
google-pasta = "0.2.0"
grpcio = "1.66.2"
h5py = "3.12.1"
huggingface-hub = "0.25.2"
idna = "3.10"
imageio = "2.35.1"
importlib-metadata = "8.5.0"
ipykernel = "6.29.5"
ipython = "8.28.0"
jedi = "0.19.1"
jinja2 = "3.1.4"
joblib = "1.4.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
keras = "3.6.0"
lazy-loader = "0.4"
libclang = "18.1.1"
markdown = "3.7"
markdown-it-py = "3.0.0"
markupsafe = "3.0.1"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
ml-dtypes = "0.5.0"
mpmath = "1.3.0"
namex = "0.0.8"
narwhals = "1.9.3"
nest-asyncio = "1.6.0"
networkx = "3.4.1"
oauthlib = "3.2.2"
opencv-python = "4.10.0.84"
opencv-python-headless = "4.10.0.84"
opt-einsum = "3.4.0"
optree = "0.13.0"
packaging = "24.1"
pandas = "2.2.3"
parso = "0.8.4"
pexpect = "4.9.0"
pickleshare = "0.7.5"
pillow = "10.4.0"
platformdirs = "4.3.6"
prompt-toolkit = "3.0.48"
protobuf = "5.28.2"
psutil = "6.0.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pyarrow = "17.0.0"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pydeck = "0.9.1"
pygments = "2.18.0"
pympler = "1.1"
python-dateutil = "2.9.0.post0"
pytz = "2024.2"
pytz-deprecation-shim = "0.1.0.post0"
pywavelets = "1.7.0"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
qudida = "0.0.4"
referencing = "0.35.1"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.9.2"
rpds-py = "0.20.0"
rsa = "4.9"
safetensors = "0.4.5"
scikit-image = "0.24.0"
scikit-learn = "1.5.2"
scipy = "1.14.1"
sentry-sdk = "2.16.0"
setproctitle = "1.3.3"
six = "1.16.0"
smmap = "5.0.1"
stack-data = "0.6.3"
streamlit = "1.39.0"
sympy = "1.13.3"
tenacity = "9.0.0"
termcolor = "2.5.0"
threadpoolctl = "3.5.0"
tifffile = "2024.9.20"
timm = "1.0.9"
toml = "0.10.2"
toolz = "1.0.0"
torch = "2.4.1"
torchvision = "0.19.1"
tornado = "6.4.1"
tqdm = "4.66.5"
traitlets = "5.14.3"
typing-extensions = "4.12.2"
tzdata = "2024.2"
tzlocal = "5.2"
urllib3 = "2.2.3"
validators = "0.34.0"
wandb = "0.18.3"
watchdog = "5.0.3"
wcwidth = "0.2.13"
werkzeug = "3.0.4"
wrapt = "1.16.0"
zipp = "3.20.2"
pytest = "^8.3.3"
pylint = "^3.3.1"
fvcore = "^0.1.5.post20221221"
llvmlite = "^0.44.0"
numba = "^0.61.0"
rembg = "^2.0.61"
onnxruntime = "^1.20.1"
matplotlib = "^3.10.0"
uvicorn = "^0.34.3"
fastapi = "^0.115.12"
python-multipart = "^0.0.20"
open-clip-torch = "^2.24.0"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"
mypy = "^1.8.0"
pytest-cov = "^6.0.0"
bandit = {extras = ["toml"], version = "^1.7.0"}
pre-commit = "^3.7.0"

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [".git", ".venv", "venv", "build", "dist", "__pycache__", "node_modules"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E203",  # whitespace before ':' (black-compatible)
    "E501",  # line too long (handled by formatter)
    "B008",  # FastAPI idiom: File(...) / Depends() in default args
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "tests/",
    ".venv/"
]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Skip assert_used check for tests

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-private = true
ignore-property-decorators = true
ignore-module = true
fail-under = 60
exclude = ["tests", "research", ".venv"]
verbose = 1
quiet = false
color = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests that take >5s (deselect with -m 'not slow')",
    "integration: integration tests requiring real model weights or external data",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
